Use derived issuer names for remote MCP auth setup
