+++ .changeset/auto-configure-remote-mcp-idp.md
@@ -1,0 +1,7 @@
+---
+"dashboard": patch
+---
+
+Auto-configure remote identity providers for custom remote MCP servers when OAuth
+metadata can be discovered, including reusing existing issuers and deriving a
+display name from the issuer URL.
